Factor out operation creation
